+++ frontend/sfem_Function.hpp
@@ -23,39 +23,13 @@
 #include "sfem_Mesh.hpp"
 #include "sfem_FunctionSpace.hpp"
 #include "sfem_glob.hpp"
+#include "sfem_Sideset.hpp"
 
 // Operator includes
 #include "sfem_Op.hpp"
 #include "sfem_OpFactory.hpp"
 
 namespace sfem {
-
-    class Sideset final {
-    public:
-        int                             read(const std::shared_ptr<Communicator>& comm, const char *path);
-        SharedBuffer<element_idx_t>     parent();
-        SharedBuffer<int16_t>           lfi();
-        static std::shared_ptr<Sideset> create_from_file(const std::shared_ptr<Communicator>& comm, const char *path);
-        ptrdiff_t                       size() const;
-        std::shared_ptr<Communicator>   comm() const;
-        int                             write(const char *path) const;
-
-        Sideset(const std::shared_ptr<Communicator>& comm, const SharedBuffer<element_idx_t> &parent, const SharedBuffer<int16_t> &lfi);
-        Sideset();
-        ~Sideset();
-
-        static std::shared_ptr<Sideset> create(const std::shared_ptr<Communicator>& comm,
-                                               const SharedBuffer<element_idx_t> &parent,
-                                               const SharedBuffer<int16_t>       &lfi);
-
-        static std::shared_ptr<Sideset> create_from_selector(
-                const std::shared_ptr<Mesh>                                         &mesh,
-                const std::function<bool(const geom_t, const geom_t, const geom_t)> &selector);
-
-    private:
-        class Impl;
-        std::unique_ptr<Impl> impl_;
-    };
 
     class NeumannConditions final : public Op {
     public:
